% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frailty_DEIE.r
\name{mediation}
\alias{mediation}
\title{mediation analysis using the transformation probit model under the case-cohort study}
\usage{
IV(ppt,data,int="TRUE",PX=NULL,tol=0.005,setting,level=lv)
}
\description{
omega.DE: the point-wise estimated direct effect obtained from Omega function (page 4)
omega.IE: the point-wise estimated indirect effect obtained from Omega function (page 4)
omega.TE: the point-wise estimated total effect obtained from Omega function (page 4)
Delta.DE: the point-wise estimated direct effect obtained from Delta (page 6)
Delta.IE: the point-wise estimated indirect effect obtained from Delta (page 6)
Delta.TE:  the point-wise estimated total effect obtained from Delta (page 6)
omega.DE.v: the point-wise estimated sd of Omega.DE
omega.IE.v: the point-wise estimated sd of Omega.IE
omega.TE.v: : the point-wise estimated sd of Omega.TE
Delta.DE.v: the estimated sd of Delta.DE
Delta.IE.v: the estimated sd of Delta.IE
Delta.TE.v: : the estimated sd of Delta.IE
log.pm: the point-wise estimated log proportion of mediation obtained by function pi_omega (page 4) 
log.pm.sd: the point-wise estimated sd of log.pm
log.pm.delta: the point-wise estimated log proportion of mediation obtained by Delta (page 4)
log.pm.delta.sd: the estimated sd of og.pm.delta
regM: the estimated regression of model (2) in page 5
coefP: the estimated regression of model (1) in page 5
}
\examples{
url <- "https://github.com/yujihchang/SurrCase/raw/main/Surrogate/Data/Wild%20Type%20Live%20Virus.RData"
destfile <- "Wild Type Live Virus.RData"
if (!file.exists(destfile)) {
  download.file(url, destfile, mode = "wb")
}
load("Wild Type Live Virus.RData")

ee = runif(30000, 0, 0.5)
data[, 1] = data[, 1] + ee  # 分開 tie
colnames(data)[1:5] = c("Ts", "D", "xi", "Z", "M")
ppt = seq(0, max(data$Ts[data$D == 1]), by = 0.01)

ans_med = mediation(ppt, data, int = "FALSE", PX = c(12, 13, 14), setting = "median")

plot(ppt, ans_med$omega.IE, col = 2, type = "l", ylim = c(-0.03, 0.03))
points(ppt, ans_med$omega.IE + 1.96 * ans_med$omega.IE.v, col = 2)
points(ppt, ans_med$omega.IE - 1.96 * ans_med$omega.IE.v, col = 2)
points(ppt, ans_med$omega.DE - 1.96 * ans_med$omega.DE.v, col = 1)
points(ppt, ans_med$omega.DE + 1.96 * ans_med$omega.DE.v, col = 1)
points(ppt, ans_med$omega.DE, col = 1)
abline(h = 0, col = 3)

}
\keyword{mediation}
