% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frailty_DEIE.r
\name{IV}
\alias{IV}
\title{IV analysis using the transformation probit model under the case-cohort study}
\usage{
IV(ppt, data, m0 = 0, m1 = 1, PX = NULL, tol = 0.005, SS = 1, mu = 0, bin = 0.5, setting, level = lv)
}
\description{
WIE: The point-wise estimated effect obtained from IV analyses through the \code{W_IE()} function. \cr
WIE.v: The point-wise estimated standard deviation of WIE. \cr
delta.IV: Estimated ratio \(\frac{b_z}{a_z}\). \cr
delta.IV.v: Estimated standard deviation of \code{delta.IV}.
}
\examples{

}
\keyword{IV}
