% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frailty_DEIE.r
\name{IV}
\alias{IV}
\title{IV analysis using the transformation probit model under the case-cohort study}
\usage{
IV(ppt,data,m0=0,m1=1,PX=NULL,tol=0.005,SS=1,mu=0,bin=0.5,setting,level=lv)
}
\description{
WIE: the point-wise estimated effect obtained from IV analyses through the W_IE() function (page 7)
WIE.v: the point-wise estimated sd of WIE
delta.IV: estimated bz/az (page 7)
delta.IV.v: estimated sd. of delta.IV
}
\examples{
IV()
}
\keyword{Frailty.DEIE}