% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frailty.r
\name{IV}
\alias{IV}
\title{IV Analysis Using the Transformation Probit Model in Case-Cohort Studie}
\usage{
IV(ppt,data,m0=0,m1=1,PX=NULL,tol=0.005,SS=1,mu=0,bin=0.5,setting,level=lv)
}
\arguments{
\item{data}{data.frame(X1,X2,D,Z)}

\item{int_theta}{initial value of theta for iteration, nonnegative values vector of length 2}

\item{tol}{maximum tolerance of change during the iteration}

\item{step}{maximum  number of the iteration}

\item{interpolation}{time can be vector or scalar}
}
\description{
Estimating the direct and indirect of the frailty model
}
\examples{
data=meta.gen(500,theta_0=0.5,theta_1=0.5,L1=0.5,L2=0.5,L3=1,b01=1,b02=0,b03=0,cc=2,dd="uniform")
P.time=seq(0,1,by=0.01)
ans=Frailty(data,P.time,int_theta=c(0.5,0.5),tol=0.01,step=50)
plot(P.time,ans$DE,type="l",ylim=c(-0.5,0.5))
points(P.time,ans$DE+ans$DE_sd,type="l",ylim=c(-0.5,0.5))
points(P.time,ans$DE-ans$DE_sd,type="l",ylim=c(-0.5,0.5))
points(P.time,ans$IE,type="l",ylim=c(-0.5,0.5),col=2)
points(P.time,ans$IE+ans$IE_sd,type="l",ylim=c(-0.5,0.5),col=2)
points(P.time,ans$IE-ans$IE_sd,type="l",ylim=c(-0.5,0.5),col=2)
legend(0,0.45,c("direct effect","indirect effect"),col=1:2,lty=1)
}
